type conversion

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,6 +1,5 @@
 #include "search_algos.h"
 
-
 /**
  * linear search - searches for a value in an array of
  * integers using the Linear search algorithm
@@ -20,8 +19,9 @@ int linear_search(int *array, size_t size, int value)
 	for (i = 0; i < size; i++)
 	{
 		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+
 		if (array[i] == value)
-			return (i);
+			return ((int)i);
 	}
 	return (-1);
 }
